Correccion de bug creador de dirigentes

+++ myapp/signals.py
@@ -8,5 +8,5 @@ from .models import Dirigente
 def eliminar_usuario_asociado(sender, instance, **kwargs):
     usuario = instance.usuario
 
-    if usuario:
+    if usuario and not Dirigente.objects.filter(usuario=usuario).exists():
         usuario.delete()
